+++ formats/utilityClass/utilitiesConfig.js
@@ -1,0 +1,174 @@
+const utilities = [
+  {
+    name: 'font-color',
+    tokenCategory: 'color',
+    tokenType: 'font',
+    cssProp: 'color',
+    variations: [''],
+    responsive: false,
+    hover: true,
+    focus: true,
+  },
+  {
+    name: 'background-color',
+    tokenCategory: 'color',
+    tokenType: 'brand',
+    cssProp: 'background-color',
+    variations: [''],
+    responsive: false,
+    hover: true,
+    focus: true,
+  },
+  {
+    name: 'border-color',
+    tokenCategory: 'color',
+    tokenType: 'brand',
+    cssProp: 'border-color',
+    variations: [''],
+    responsive: false,
+    hover: true,
+    focus: true,
+  },
+  {
+    name: 'font-size',
+    tokenCategory: 'size',
+    tokenType: 'font',
+    cssProp: 'font-size',
+    variations: [''],
+    responsive: true,
+    hover: true,
+    focus: false,
+  },
+  {
+    name: 'border-radius',
+    abbreviation: 'br',
+    tokenCategory: 'size',
+    tokenType: 'border-radius',
+    cssProp: 'border-radius',
+    variations: ['', 'top-left', 'top-right', 'bottom-right', 'bottom-left'],
+    responsive: true,
+    hover: false,
+    focus: false,
+  },
+  {
+    name: 'margin',
+    abbreviation: 'm',
+    tokenCategory: 'size',
+    tokenType: 'spacing',
+    cssProp: 'margin',
+    variations: ['', 'top', 'right', 'bottom', 'left', 'h', 'v'],
+    responsive: true,
+    hover: false,
+    focus: false,
+  },
+  {
+    name: 'padding',
+    abbreviation: 'p',
+    tokenCategory: 'size',
+    tokenType: 'spacing',
+    cssProp: 'padding',
+    variations: ['', 'top', 'right', 'bottom', 'left', 'h', 'v'],
+    responsive: true,
+    hover: false,
+    focus: false,
+  },
+  {
+    name: 'width',
+    abbreviation: 'w',
+    tokenCategory: 'size',
+    tokenType: 'width',
+    cssProp: 'width',
+    variations: [''],
+    responsive: true,
+    hover: false,
+    focus: false,
+  },
+  {
+    name: 'maxWidth',
+    abbreviation: 'mw',
+    tokenCategory: 'size',
+    tokenType: 'width',
+    cssProp: 'max-width',
+    variations: [''],
+    responsive: true,
+    hover: false,
+    focus: false,
+  },
+  {
+    name: 'minWidth',
+    abbreviation: 'minw',
+    tokenCategory: 'size',
+    tokenType: 'width',
+    cssProp: 'min-width',
+    variations: [''],
+    responsive: true,
+    hover: false,
+    focus: false,
+  },
+  {
+    name: 'height',
+    abbreviation: 'h',
+    tokenCategory: 'size',
+    tokenType: 'height',
+    cssProp: 'height',
+    variations: [''],
+    responsive: true,
+    hover: false,
+    focus: false,
+  },
+  {
+    name: 'maxHeight',
+    abbreviation: 'mh',
+    tokenCategory: 'size',
+    tokenType: 'height',
+    cssProp: 'max-height',
+    variations: [''],
+    responsive: true,
+    focus: false,
+  },
+  {
+    name: 'minHeight',
+    abbreviation: 'minh',
+    tokenCategory: 'size',
+    tokenType: 'height',
+    cssProp: 'min-height',
+    variations: [''],
+    responsive: true,
+    hover: false,
+    focus: false,
+  },
+  {
+    name: 'border-width',
+    tokenCategory: 'size',
+    tokenType: 'border',
+    cssProp: 'border',
+    variations: ['', 'top', 'right', 'bottom', 'left', 'h', 'v'],
+    responsive: true,
+    hover: false,
+    focus: true,
+  },
+  {
+    name: 'font-weight',
+    tokenCategory: 'size',
+    tokenType: 'font-weight',
+    cssProp: 'font-weight',
+    variations: [''],
+    responsive: true,
+    hover: false,
+    focus: false,
+  },
+  {
+    name: 'box-shadow',
+    abbreviation: 'shadow',
+    tokenCategory: 'size',
+    tokenType: 'box-shadow',
+    cssProp: 'box-shadow',
+    variations: [''],
+    responsive: true,
+    hover: true,
+    hover: false,
+    focus: true,
+  }
+];
+
+module.exports = utilities;
